+++ app/build.gradle
@@ -42,7 +42,6 @@ dependencies {
     }
 
     // include the extensions module in classpath to be able to use it with Crate launched from within intellij
-    runtime project(':extensions:ssl-impl')
     runtime project(':extensions:functions')
     runtime project(':extensions:lang-js')
 
@@ -275,20 +274,6 @@ jar {
     from fileTree(buildDir).matching { include 'META-INF/services/*' }
 }
 
-task collectExtensionsModules(
-    dependsOn: [':extensions:ssl-impl:jar']) {
-    // this task collects the extensions module jar into a single folder
-    // named extensions from which the files are taken for the installDist task
-    doLast {
-        copy {
-            from(project(':extensions:ssl-impl').tasks.jar.archivePath)
-            into 'extensions'
-        }
-    }
-}
-collectExtensionsModules.outputs.dir('extensions')
-
-
 task downloadPlugins(
     dependsOn: ['downloadAdminUI',
                 ':plugins:es-repository-hdfs:jar',
@@ -359,7 +344,7 @@ task downloadAdminUI {
 }
 
 
-processResources.dependsOn(downloadPlugins, collectExtensionsModules)
+processResources.dependsOn(downloadPlugins)
 
 task(runDebug, dependsOn: 'classes', type: JavaExec) {
     main = 'io.crate.bootstrap.CrateDB'
@@ -393,8 +378,7 @@ test {
 }
 
 
-clean.dependsOn(['cleanDownloadPlugins',
-                 'cleanCollectExtensionsModules'])
+clean.dependsOn(['cleanDownloadPlugins'])
 
 
 def extractTopfolder(File src, String trg) {
