--- extensions/ssl-impl/build.gradle
@@ -1,27 +1,0 @@
-apply from: "$rootDir/extensions/extensionModule.gradle"
-
-archivesBaseName = 'ssl-impl'
-group = 'io.crate'
-description = 'SSL Encryption Implementation for CrateDB'
-
-dependencies {
-    implementation project(':libs:shared')
-    implementation project(':server')
-    implementation "io.netty:netty-handler:${versions.netty4}"
-    implementation "org.apache.logging.log4j:log4j-api:${versions.log4j2}"
-
-    testImplementation project(':server')
-    testImplementation project(path: ':server', configuration: 'testOutput')
-    testImplementation project(path: ':libs:dex', configuration: 'testOutput')
-    testImplementation "org.postgresql:postgresql:${versions.jdbc}"
-    testImplementation "org.mockito:mockito-core:${versions.mockito}"
-    testImplementation 'org.skyscreamer:jsonassert:1.3.0'
-    testImplementation "org.apache.lucene:lucene-test-framework:${versions.lucene}"
-    testImplementation "com.carrotsearch.randomizedtesting:randomizedtesting-runner:${versions.randomizedrunner}"
-    testImplementation "org.hamcrest:hamcrest:${versions.hamcrest}"
-    testImplementation "org.junit.jupiter:junit-jupiter:${versions.junit5}"
-    testImplementation "junit:junit:${versions.junit}"
-    testRuntimeOnly("org.junit.vintage:junit-vintage-engine") {
-        because 'allows JUnit 3 and JUnit 4 tests to run'
-    }
-}
