--- extensions/ssl-impl/src/main/resources/META-INF/services/io.crate.protocols.ssl.SslExtension
@@ -1,1 +1,0 @@
-io.crate.plugin.EnterpriseSslExtension
