+++ server/src/main/java/io/crate/auth/AuthenticationModule.java
@@ -22,10 +22,9 @@
 
 package io.crate.auth;
 
-import io.crate.auth.AlwaysOKAuthentication;
-import io.crate.auth.AuthSettings;
-import io.crate.auth.Authentication;
-import io.crate.auth.HostBasedAuthentication;
+import io.crate.protocols.ssl.SslContextProvider;
+import io.crate.protocols.ssl.SslContextProviderImpl;
+
 import org.elasticsearch.common.inject.AbstractModule;
 import org.elasticsearch.common.settings.Settings;
 
@@ -45,5 +44,6 @@ public class AuthenticationModule extends AbstractModule {
             bind(Authentication.class).to(AlwaysOKAuthentication.class);
         }
         bind(AuthenticationHttpAuthHandlerRegistry.class).asEagerSingleton();
+        bind(SslContextProvider.class).to(SslContextProviderImpl.class).asEagerSingleton();
     }
 }
