+++ server/src/test/java/io/crate/auth/AuthenticationWithSSLIntegrationTest.java
