+++ server/src/test/java/io/crate/protocols/http/HttpsTransportKeyStoreReloadIntegrationTest.java
@@ -20,16 +20,18 @@
  * agreement.
  */
 
-package io.crate.protocols.postgres;
+package io.crate.protocols.http;
 
-import io.crate.integrationtests.SQLTransportIntegrationTest;
+import io.crate.integrationtests.SQLHttpIntegrationTest;
 import io.crate.protocols.ssl.SslConfigSettings;
-import io.crate.testing.SQLResponse;
 import io.crate.testing.UseJdbc;
 import io.netty.handler.ssl.util.SelfSignedCertificate;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.util.EntityUtils;
 import org.elasticsearch.common.CheckedConsumer;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.test.ESIntegTestCase;
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
@@ -40,24 +42,30 @@ import java.security.KeyStore;
 import java.security.cert.Certificate;
 import java.util.concurrent.TimeUnit;
 
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.Matchers.not;
+import static org.hamcrest.Matchers.nullValue;
 
-@UseJdbc(value = 1)
+@UseJdbc(value = 0)
 @ESIntegTestCase.ClusterScope(numDataNodes = 1, numClientNodes = 0, supportsDedicatedMasters = false)
-public class SslReqHandlerIntegrationTest extends SQLTransportIntegrationTest {
+public class HttpsTransportKeyStoreReloadIntegrationTest extends SQLHttpIntegrationTest {
 
     private final static char[] EMPTY_PASS = new char[]{};
 
     private static SelfSignedCertificate trustedCert;
     private static File keyStoreFile;
+    private static File trustStoreFile;
 
-    public SslReqHandlerIntegrationTest() {
+    public HttpsTransportKeyStoreReloadIntegrationTest() {
         super(true);
     }
 
     @BeforeClass
     public static void beforeTest() throws Exception {
         trustedCert = new SelfSignedCertificate();
+
         keyStoreFile = createTempFile().toFile();
+        trustStoreFile = createTempFile().toFile();
 
         updateKeyStore(
             keyStoreFile,
@@ -65,10 +73,21 @@ public class SslReqHandlerIntegrationTest extends SQLTransportIntegrationTest {
                 SelfSignedCertificate fakeCert = new SelfSignedCertificate();
                 store.setKeyEntry(
                     "key",
-                    trustedCert.key(),
+                    fakeCert.key(),
                     EMPTY_PASS,
                     new Certificate[]{fakeCert.cert()});
             });
+
+        updateKeyStore(
+            trustStoreFile,
+            keyStore -> keyStore.setCertificateEntry("cert", trustedCert.cert()));
+
+        System.setProperty("javax.net.ssl.trustStore", trustStoreFile.getAbsolutePath());
+    }
+
+    @AfterClass
+    public static void afterTest() {
+        System.clearProperty("javax.net.ssl.trustStore");
     }
 
     private static void updateKeyStore(File keyStoreFile, CheckedConsumer<KeyStore, Exception> consumer)
@@ -80,24 +99,26 @@ public class SslReqHandlerIntegrationTest extends SQLTransportIntegrationTest {
             keyStore.load(null);
         }
         consumer.accept(keyStore);
-        keyStore.store(new FileOutputStream(keyStoreFile), EMPTY_PASS);
+        try (var fs = new FileOutputStream(keyStoreFile)) {
+            keyStore.store(fs, EMPTY_PASS);
+        }
     }
 
     @Override
     protected Settings nodeSettings(int nodeOrdinal) {
         return Settings.builder()
             .put(super.nodeSettings(nodeOrdinal))
-            .put(SslConfigSettings.SSL_PSQL_ENABLED.getKey(), true)
+            .put(SslConfigSettings.SSL_HTTP_ENABLED.getKey(), true)
             .put(SslConfigSettings.SSL_KEYSTORE_FILEPATH.getKey(), keyStoreFile.getAbsolutePath())
+            .put(SslConfigSettings.SSL_TRUSTSTORE_FILEPATH.getKey(), trustStoreFile.getAbsolutePath())
             .put(SslConfigSettings.SSL_RESOURCE_POLL_INTERVAL.getKey(), "2s")
             .build();
     }
 
     @Test
     public void testReloadSslContextOnKeyStoreChange() throws Throwable {
-        try {
-            execute("select name from sys.nodes");
-            fail("Unknown certificate in the certificate chain");
+        try (CloseableHttpResponse ignored = post("{\"stmt\": \"select 'sslWorks'\"}")) {
+            fail("Certificates in the trusted store signed with unknown key");
         } catch (Exception ignored) {
         }
 
@@ -110,9 +131,12 @@ public class SslReqHandlerIntegrationTest extends SQLTransportIntegrationTest {
                 new Certificate[]{trustedCert.cert()}));
 
         assertBusy(() -> {
-            try {
-                SQLResponse response = execute("select name from sys.nodes");
-                assertEquals(1, response.rowCount());
+            try (CloseableHttpResponse response = post("{\"stmt\": \"select 'sslWorks'\"}")) {
+                assertThat(response, not(nullValue()));
+                assertEquals(200, response.getStatusLine().getStatusCode());
+                String result = EntityUtils.toString(response.getEntity());
+                assertThat(result, containsString("\"rowcount\":1"));
+                assertThat(result, containsString("sslWorks"));
             } catch (Exception e) {
                 fail(e.getMessage());
             }
