+++ settings.gradle
@@ -20,7 +20,6 @@ include 'libs:shared',
         'blackbox',
         'extensions:jmx-monitoring',
         'extensions:lang-js',
-        'extensions:ssl-impl',
         'extensions:functions',
         'benchmarks'
